+++ MovieListFramework/Media API/HTTPClient.swift
@@ -1,0 +1,17 @@
+//
+//  HTTPClient.swift
+//  MovieListFramework
+//
+//  Created by macbook on 08/09/2023.
+//
+
+import Foundation
+
+public enum HTTPClientResult {
+    case success(Data, HTTPURLResponse)
+    case failure(Error)
+}
+
+public protocol HTTPClient {
+    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
+}
