+++ MovieListFramework/Media API/RemoteMediaLoader.swift
@@ -7,15 +7,6 @@
 
 import Foundation
 
-public enum HTTPClientResult {
-    case success(Data, HTTPURLResponse)
-    case failure(Error)
-}
-
-public protocol HTTPClient {
-    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
-}
-
 public final class RemoteMediaLoader {
     private let url: URL
     private let client: HTTPClient
